embed

+++ ml_utility_loss/loss_learning/estimator/wrapper.py
@@ -73,8 +73,8 @@ class MLUtilityTrainer:
             samples = samples.type(train.dtype)
 
         if self.model.adapter.embedding and train.dim() < 4:
-            train = self.model.adapter.embedding(train)
-            test = self.model.adapter.embedding(test)
+            train = self.model.adapter.embedding(train.to(torch.int))
+            test = self.model.adapter.embedding(test.to(torch.int))
 
         assert train.dim() == samples.dim() and train.shape[0] == samples.shape[0] == 1 and train.shape[-1] == samples.shape[-1], f"Mismatching shapes. train {train.shape}, samples {samples.shape}"
